refactor(bundler): move generate_submodule_attributes_with_exclusions to namespace_manager
Extract 310-line function from bundler.rs to new namespace_manager.rs module
as part of Phase 7 token reduction initiative. This achieves significant
file size reduction while preserving full functionality.

- Create new namespace_manager.rs module with complete function implementation
- Update bundler.rs imports and call sites to use extracted function
- Make helper methods create_namespace_module and create_dotted_attribute_assignment public
- Maintain all original logic for namespace hierarchy analysis and module assignments
- Preserve deterministic ordering and complex exclusion handling

All tests pass (89 tests) confirming no functional regressions.
